requirements.txt
pypdf
google-generativeai
argparse
load_dotenv
tqdm
